requirements.txt
# To install these libraries run the following: pip install -r requirements.txt
#some of these libraries are not necessary
albumentations==1.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
deep-sort-realtime==1.3.2
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
idna==3.7
imageio==2.36.0
iso8601==2.1.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lap==0.5.12
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.51
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
serial==0.0.97
setuptools==75.1.0
shapely==2.0.7
six==1.16.0
solutions==0.0.2
sympy==1.13.3
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.44
ultralytics-thop==2.0.11
urllib3==2.2.3
wheel==0.44.0
